Added heft policy state
