+++ simulator/opendc-experiments/opendc-experiments-allocateam/src/main/kotlin/org/opendc/experiments/allocateam/experiment/Portfolios.kt
@@ -10,21 +10,21 @@ import org.opendc.experiments.sc20.experiment.model.Workload
 public class AllocateamPortfolio(parent: Experiment, id: Int) : Portfolio(parent, id, "allocateam") {
     override val topologies: List<Topology> = listOf(
         Topology("small"),
-        Topology("medium"),
-        Topology("large")
+//        Topology("medium"),
+//        Topology("large")
     )
 
     override val workloads: List<Workload> = listOf(
-        Workload("shell", 1.0),
+//        Workload("shell", 1.0),
         Workload("askalon_ee", 1.0),
-        Workload("spec_trace-2", 1.0),
+//        Workload("spec_trace-2", 1.0),
     )
 
     override val resourceAllocationPolicies: List<String> = listOf(
-        "min-min",
-        "max-min",
-        "round-robin",
-        "lottery",
+//        "min-min",
+//        "max-min",
+//        "round-robin",
+//        "lottery",
         "heft"
     )
